resolving image management issues

+++ models/user.js
@@ -72,7 +72,7 @@ const User = sequelize.define('user',
       afterUpdate: (user) => {
         // If pic is changed check if that pic is used elsewhere, if not - delete all copies
         let dir = path.join(__dirname, '../images/' + user.id + '/');
-        if (user.changed('profilePic')) {
+        if (user.changed('profilePic') && user.previous('profilePic') !== null) {
           if (user.previous('profilePic').includes(process.env.IMAGE_PATH) && Object.values(user.dataValues).indexOf(user.previous('profilePic')) < 0) {
             fs.unlink(dir + user.previous('profilePic').split('/' + user.id + '/')[1].split('.')[0] + '_THUMB.' + user.previous('profilePic').split('.')[1], error => debug(error));
             fs.unlink(dir + user.previous('profilePic').split('/' + user.id + '/')[1].split('.')[0] + '_SMALL.' + user.previous('profilePic').split('.')[1], error => debug(error));
@@ -80,19 +80,19 @@ const User = sequelize.define('user',
             fs.unlink(dir + user.previous('profilePic').split('/' + user.id + '/')[1].split('.')[0] + '_LARGE.' + user.previous('profilePic').split('.')[1], error => debug(error));
           }
         }
-        if (user.changed('pic1')) {
+        if (user.changed('pic1') && user.previous('pic1') !== null) {
           if (Object.values(user.dataValues).indexOf(user.previous('pic1')) < 0) {
             fs.unlink(dir + user.previous('pic1').split('/' + user.id + '/')[1].split('.')[0] + '_SMALL.' + user.previous('pic1').split('.')[1], error => debug(error));
             fs.unlink(dir + user.previous('pic1').split('/' + user.id + '/')[1].split('.')[0] + '_LARGE.' + user.previous('pic1').split('.')[1], error => debug(error));
           }
         }
-        if (user.changed('pic2')) {
+        if (user.changed('pic2') && user.previous('pic2') !== null) {
           if (Object.values(user.dataValues).indexOf(user.previous('pic2')) < 0) {
             fs.unlink(dir + user.previous('pic2').split('/' + user.id + '/')[1].split('.')[0] + '_SMALL.' + user.previous('pic2').split('.')[1], error => debug(error));
             fs.unlink(dir + user.previous('pic2').split('/' + user.id + '/')[1].split('.')[0] + '_LARGE.' + user.previous('pic2').split('.')[1], error => debug(error));
           }
         }
-        if (user.changed('pic3')) {
+        if (user.changed('pic3') && user.previous('pic3') !== null) {
           if (Object.values(user.dataValues).indexOf(user.previous('pic3')) < 0) {
             fs.unlink(dir + user.previous('pic3').split('/' + user.id + '/')[1].split('.')[0] + '_SMALL.' + user.previous('pic3').split('.')[1], error => debug(error));
             fs.unlink(dir + user.previous('pic3').split('/' + user.id + '/')[1].split('.')[0] + '_LARGE.' + user.previous('pic3').split('.')[1], error => debug(error));
